further tweaking, still not working

+++ release.sh
@@ -1,6 +1,20 @@
 #!/bin/bash
 echo "Creating a new release"
 
+while getopts "h?d" opt; do
+    case "$opt" in
+      h|\?)
+        echo "create and publish a new release. Pass the option -d to run without restarting and publishing in th docker containers"
+        exit 0
+        ;;
+      d)  devmode=1
+        echo "running in devmode"
+        ;;
+      f)  output_file=$OPTARG
+        ;;
+    esac
+done
+
 migration_version=$(cat package.json  | jq -r '.dependencies."@daostack/migration"')
 docker_compose_migration_version=$(cat docker-compose.yml | grep daostack/migration | cut -d ":" -f 3 | sed "s/'//")
 package_version=$(cat package.json | jq -r '.version')
@@ -12,16 +26,24 @@ if [[ $docker_compose_migration_version != $migration_version ]]; then
   exit
 fi
 
-echo "Starting fresh docker containers..."
-docker-compose down -v
-docker-compose up -d
+# if [[ $devmode != 1 ]]; then
+  echo "Starting fresh docker containers..."
+  docker-compose down -v
+  docker-compose up -d
+# fi
+
+# clean up local environment
+rm migration.json
+rm subgraph.yaml
+rm scheme.graphql
+rm -r daos/private/*
 
 echo "waiting for ganache to start"
 while [[ "$(curl -s -o /dev/null -w ''%{http_code}'' 127.0.0.1:8545)" != "400" ]]; do sleep 5; done
 
+
 # copy the migration file from the migration repo
 echo "deploying ethereum contracts and doing transactions...."
-rm migration.json
 docker-compose exec  ganache cat migration.json > migration.json
 npm run deploy-ethereum
 
@@ -37,36 +59,39 @@ while [[ $(curl --silent -H "Content-Type: application/json" -d '{"query":"{ sub
 echo "subgraph is synced"
 
 
-# commit the ganache image
-container_id=$(docker ps  -f "name=ganache" -l -q)
-image_name=daostack/test-env
-echo "docker commit $container_id $image_name:$image_version"
-docker commit $container_id $image_name:$image_version
-echo "docker push $image_name:$image_version"
-docker push $image_name:$image_version
+if [[ $devmode != 1 ]]; then
+  echo "published new docker images"
+  # commit the ganache image
+  container_id=$(docker ps  -f "name=ganache" -l -q)
+  image_name=daostack/test-env
+  echo "docker commit $container_id $image_name:$image_version"
+  docker commit $container_id $image_name:$image_version
+  echo "docker push $image_name:$image_version"
+  docker push $image_name:$image_version
 
-# commit the postgres image
-container_id=$(docker ps  -f "name=postgres" -l -q)
-image_name=daostack/subgraph-postgres
-echo "docker commit $container_id $image_name:$image_version"
-docker commit $container_id $image_name:$image_version
-echo "docker push $image_name:$image_version"
-docker push $image_name:$image_version
+  # commit the postgres image
+  container_id=$(docker ps  -f "name=postgres" -l -q)
+  image_name=daostack/subgraph-postgres
+  echo "docker commit $container_id $image_name:$image_version"
+  docker commit $container_id $image_name:$image_version
+  echo "docker push $image_name:$image_version"
+  docker push $image_name:$image_version
 
-# commit the ipfs  image
-container_id=$(docker ps  -f "name=ipfs" -l -q)
-image_name=daostack/subgraph-ipfs
-echo "docker commit $container_id $image_name:$image_version"
-docker commit $container_id $image_name:$image_version
-echo "docker push $image_name:$image_version"
-docker push $image_name:$image_version
+  # commit the ipfs  image
+  container_id=$(docker ps  -f "name=ipfs" -l -q)
+  image_name=daostack/subgraph-ipfs
+  echo "docker commit $container_id $image_name:$image_version"
+  docker commit $container_id $image_name:$image_version
+  echo "docker push $image_name:$image_version"
+  docker push $image_name:$image_version
 
-docker-compose down -v
-# tag on github
-echo "create tag ${image_version}"
-git tag -a $image_version -m "Release of version $image_name:$image_version"
-git push --tags
-# done
+  docker-compose down -v
+  # tag on github
+  echo "create tag ${image_version}"
+  git tag -a $image_version -m "Release of version $image_name:$image_version"
+  git push --tags
+  # done
 
-doker-compose down -v
-echo "Done!"
+  doker-compose down -v
+  echo "Done!"
+fi
